* tools/tsstdin.c: fixed check for possible negative microsecond value. clib2 uses unsigned integers for struct timeval only.

+++ tools/tsstdin.c
@@ -10,7 +10,7 @@
   do {                                                                        \
     (result)->tv_sec = (a)->tv_sec - (b)->tv_sec;                             \
     (result)->tv_usec = (a)->tv_usec - (b)->tv_usec;                          \
-    if ((result)->tv_usec < 0) {                                              \
+    if ((long)((result)->tv_usec) < 0) {                                      \
       --(result)->tv_sec;                                                     \
       (result)->tv_usec += 1000000;                                           \
     }                                                                         \
